conformer-transformer/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
appnope==0.1.3
arrow==1.2.3
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
backcall==0.2.0
beautifulsoup4==4.12.0
blessed==1.20.0
certifi @ file:///croot/certifi_1671487769961/work/certifi
charset-normalizer==3.1.0
click==8.1.3
cmake==3.26.1
colorama==0.4.6
comm==0.1.3
contourpy==1.0.7
croniter==1.3.8
cycler==0.11.0
dateutils==0.6.12
debugpy==1.6.6
decorator==5.1.1
deepdiff==6.3.0
dnspython==2.3.0
email-validator==1.3.1
entmax==1.1
executing==1.2.0
fastapi==0.88.0
filelock==3.10.7
fonttools==4.39.3
frozenlist==1.3.3
fsspec==2023.3.0
h11==0.14.0
h5py==3.8.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
huggingface-hub==0.12.1
idna==3.4
inquirer==3.1.3
ipykernel==6.22.0
ipython==8.12.0
ipywidgets==8.0.6
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonargparse==3.13.1
jupyter_client==8.1.0
jupyter_core==5.3.0
jupyterlab-widgets==3.0.7
kiwisolver==1.4.4
lightning==2.0.1
lightning-cloud==0.5.32
lightning-utilities==0.8.0
lit==16.0.0
lxml==4.9.2
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
nest-asyncio==1.5.6
networkx==3.0
nltk==3.8.1
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
ordered-set==4.1.0
orjson==3.8.9
packaging==23.0
pandas==1.5.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.2.0
portalocker==2.7.0
prompt-toolkit==3.0.38
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==1.10.7
pyDeprecate==0.3.2
Pygments==2.14.0
PyJWT==2.6.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
python-editor==1.0.4
python-multipart==0.0.6
pytorch-lightning==1.9.5
pytz==2023.3
PyYAML==6.0
pyzmq==25.0.2
readchar==4.0.5
regex==2023.3.23
requests==2.28.2
rfc3986==1.5.0
rich==13.3.3
sacrebleu==2.3.1
scipy==1.10.1
seaborn==0.12.2
sentencepiece==0.1.97
six==1.16.0
sniffio==1.3.0
soupsieve==2.4
stack-data==0.6.2
starlette==0.22.0
starsessions==1.3.0
sympy==1.11.1
tabulate==0.9.0
tokenizers==0.13.3
torch==2.0.0
torchmetrics==0.8.2
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.28.1
triton==2.0.0
typing_extensions==4.5.0
ujson==5.7.0
unbabel-comet==2.0.1
urllib3==1.26.15
uvicorn==0.21.1
uvloop==0.17.0
watchfiles==0.19.0
wcwidth==0.2.6
websocket-client==1.5.1
websockets==10.4
widgetsnbextension==4.0.7
yarl==1.8.2
